pyproject.toml
[tool.poetry]
name = "supportbot"
version = "0.1.0"
description = ""
authors = ["uiolo <vladyan198@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
redis = {extras = ["hiredis"], version = "^5.0.3"}
requests = "^2.31.0"
arq = "^0.25.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
